[refactor] change functions name
