+++ audio_visualizer/src/main/java/com/miller198/audiovisualizer/soundeffect/SoundEffectWaveStroke.kt
@@ -1,29 +1,21 @@
 package com.miller198.audiovisualizer.soundeffect
 
-import androidx.compose.animation.core.LinearOutSlowInEasing
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Rect
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.ClipOp
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Path
-import androidx.compose.ui.graphics.drawscope.clipPath
+import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.layout.onSizeChanged
-import com.miller198.audiovisualizer.soundeffect.DrawSoundEffectConfigs.animatedGradientRadius
-import com.miller198.audiovisualizer.soundeffect.DrawSoundEffectConfigs.gradientConfig
-import com.miller198.audiovisualizer.soundeffect.DrawSoundEffectConfigs.onCanvasSizeChanged
+import com.miller198.audiovisualizer.soundeffect.SoundEffectConfigs.onCanvasSizeChanged
 
 /**
- * Composable function that draws a smooth radial waveform (filled area).
+ * Composable function that draws a smooth radial waveform (outline only).
  *
  * This effect visualizes the waveform in a circular path using Catmull-Rom splines.
  *
@@ -32,7 +24,7 @@ import com.miller198.audiovisualizer.soundeffect.DrawSoundEffectConfigs.onCanvas
  * @param modifier Modifier to apply to the Canvas layout.
  */
 @Composable
-internal fun DrawSoundWaveFill(
+internal fun SoundEffectWaveStroke(
     audioData: List<Float>,
     color: Color,
     modifier: Modifier = Modifier,
@@ -40,11 +32,8 @@ internal fun DrawSoundWaveFill(
     /** The radius from the center to the start of the bars */
     var adjustedRadius by remember { mutableFloatStateOf(0f) }
 
-    /**  The maximum possible bar height based on canvas size */
+    /** The maximum possible bar height based on canvas size */
     var maxEffectHeight by remember { mutableFloatStateOf(0f) }
-
-    /** Path used to define the circular center hole (to clip out) */
-    var holePath by remember { mutableStateOf(Path()) }
 
     /** Angle between each bar in the 360° circle */
     val angleStep = 360f / audioData.size
@@ -52,10 +41,7 @@ internal fun DrawSoundWaveFill(
     /** The path that represents the full waveform */
     val path = Path()
 
-    /** Animated gradient radius for visual effect */
-    val animatedGradientRadius = animatedGradientRadius(LinearOutSlowInEasing)
-
-    /** Main canvas for drawing the sound wave */
+    /** Main canvas for drawing the sound bars */
     Canvas(
         modifier = modifier
             .fillMaxSize()
@@ -67,14 +53,6 @@ internal fun DrawSoundWaveFill(
                     onRadiusCalculated = { adjustedRadius = it },
                     onMaxEffectHeightCalculated = { maxEffectHeight = it }
                 )
-                holePath = holePath.apply {
-                    addOval(
-                        Rect(
-                            center = Offset(x = canvasSize.width / 2f, y = canvasSize.height / 2f),
-                            radius = adjustedRadius + 1f
-                        )
-                    )
-                }
             }
     ) {
         val width = size.width
@@ -96,24 +74,10 @@ internal fun DrawSoundWaveFill(
         // Apply Catmull-Rom spline for smooth curve interpolation
         path.catmullRomSpline(points)
 
-        // Clip the central hole area and draw the wave around it
-        clipPath(holePath, clipOp = ClipOp.Difference) {
-            if (gradientConfig.useGradient) {
-                drawPath(
-                    path = path,
-                    brush = Brush.radialGradient(
-                        colors = listOf(gradientConfig.color, color),
-                        center = Offset(width / 2, height / 2),
-                        radius = (adjustedRadius + maxEffectHeight).coerceAtLeast(0.01f)
-                                * animatedGradientRadius
-                    )
-                )
-            } else {
-                drawPath(
-                    path = path,
-                    color = color
-                )
-            }
-        }
+        drawPath(
+            path = path,
+            color = color,
+            style = Stroke(width = DrawSoundWaveStrokeConstants.STROKE_WIDTH)
+        )
     }
 }
